Minor text fixes

+++ Libraries/HyTech_CAN/ACU_shunt_measurements.h
@@ -23,9 +23,9 @@ class ACU_shunt_measurements {
     inline uint16_t get_ts_out_filtered()         const {return ts_out_filtered;}
 
     // Setters
-    inline void set_shunt_current(float shunt_current) {this->shunt_current = (uint16_t) (shunt_current);}
-    inline void set_pack_filtered(float pack_filtered) {this->pack_filtered = (uint16_t) (pack_filtered);}
-    inline void set_ts_out_filtered(float ts_out_filtered) {this->ts_out_filtered = (uint16_t) (ts_out_filtered);}
+    inline void set_shunt_current(uint16_t shunt_current) {this->shunt_current = (uint16_t) (shunt_current);}
+    inline void set_pack_filtered(uint16_t pack_filtered) {this->pack_filtered = (uint16_t) (pack_filtered);}
+    inline void set_ts_out_filtered(uint16_t ts_out_filtered) {this->ts_out_filtered = (uint16_t) (ts_out_filtered);}
 
     // Member variables
     private:
